MLMProject/requirements.txt
psycopg2-binary==2.8.6
Pillow==8.0.0
celery==4.4.5
django-celery-email==3.0.0
django-celery==3.3.1
